+++ packages/iov-bcp-types/src/atomicswaptypes.ts
@@ -1,3 +1,4 @@
+import { As } from "type-tagger";
 import { Stream } from "xstream";
 
 import { BcpConnection } from "./connection";
@@ -13,7 +14,12 @@ export interface SwapData {
   readonly id: SwapIdBytes; // this is used as an unique identitier to locate the swap
   readonly sender: Address;
   readonly recipient: Address;
-  readonly hashlock: Uint8Array; // this is the hash, whose preimage releases the swap
+  /**
+   * The hash, whose preimage releases the atomic swap.
+   *
+   * Until we have a way to specify the hashing algirithm, this is SHA256.
+   */
+  readonly hash: Uint8Array;
   readonly amounts: ReadonlyArray<Amount>;
   readonly timeout: number; // blockheight where the swap expires (TODO: alternatively support Date?)
   readonly memo?: string;
@@ -25,11 +31,13 @@ export interface OpenSwap {
   readonly data: SwapData;
 }
 
+export type Preimage = Uint8Array & As<"preimage">;
+
 // ClosedSwap is returned once the swap has been claimed, exposing the preimage that was used to claim it
 export interface ClaimedSwap {
   readonly kind: SwapState.Claimed;
   readonly data: SwapData;
-  readonly preimage: Uint8Array;
+  readonly preimage: Preimage;
 }
 
 // ExpiredSwap is an offer that timed out without being claimed