+++ packages/iov-bcp-types/src/transactions.ts
@@ -2,6 +2,8 @@ import { As } from "type-tagger";
 
 import { Int53 } from "@iov/encoding";
 
+import { Preimage } from "./atomicswaptypes";
+
 export enum Algorithm {
   Ed25519 = "ed25519",
   Secp256k1 = "secp256k1",
@@ -207,28 +209,25 @@ export interface SendTransaction extends UnsignedTransaction {
   readonly memo?: string;
 }
 
+/** A swap offer or a counter offer */
 export interface SwapOfferTransaction extends UnsignedTransaction {
   readonly kind: "bcp/swap_offer";
   readonly amounts: ReadonlyArray<Amount>;
   readonly recipient: Address;
   /** absolute block height at which the offer times out */
   readonly timeout: number;
-  readonly preimage: Uint8Array;
-}
-
-export interface SwapCounterTransaction extends UnsignedTransaction {
-  readonly kind: "bcp/swap_counter";
-  readonly amounts: ReadonlyArray<Amount>;
-  readonly recipient: Address;
-  /** absolute block height at which the counter offer times out */
-  readonly timeout: number;
-  readonly hashCode: Uint8Array; // pulled from the offer transaction
+  /**
+   * Locally calculated hash of the preimage.
+   *
+   * This is a SHA256 hash until we have a way to specifiy the hashing algorithm.
+   */
+  readonly hash: Uint8Array;
   readonly memo?: string;
 }
 
 export interface SwapClaimTransaction extends UnsignedTransaction {
   readonly kind: "bcp/swap_claim";
-  readonly preimage: Uint8Array;
+  readonly preimage: Preimage;
   readonly swapId: SwapIdBytes; // pulled from the offer transaction
 }
 
@@ -245,12 +244,6 @@ export function isSwapOfferTransaction(
   transaction: UnsignedTransaction,
 ): transaction is SwapOfferTransaction {
   return (transaction as SwapOfferTransaction).kind === "bcp/swap_offer";
-}
-
-export function isSwapCounterTransaction(
-  transaction: UnsignedTransaction,
-): transaction is SwapCounterTransaction {
-  return (transaction as SwapCounterTransaction).kind === "bcp/swap_counter";
 }
 
 export function isSwapClaimTransaction(