+++ packages/iov-bcp-types/types/index.d.ts
@@ -1,4 +1,5 @@
-export * from "./atomicswap";
+export * from "./atomicswaptypes";
+export { AtomicSwapHelpers } from "./atomicswaphelpers";
 export { AtomicSwapMerger } from "./atomicswapmerger";
 export * from "./chainconnector";
 export * from "./codec";