+++ packages/iov-bns/src/bnscodec.spec.ts
@@ -11,7 +11,7 @@ import {
   signedTxBin,
   signedTxJson,
   swapClaimTxJson,
-  swapCounterTxJson,
+  swapOfferTxJson,
   swapTimeoutTxJson,
 } from "./testdata.spec";
 
@@ -54,7 +54,7 @@ describe("bnscodec", () => {
     const transactionsToBeVerified: ReadonlyArray<SignedTransaction> = [
       signedTxJson,
       randomTxJson,
-      swapCounterTxJson,
+      swapOfferTxJson,
       swapClaimTxJson,
       swapTimeoutTxJson,
     ];
@@ -62,7 +62,9 @@ describe("bnscodec", () => {
     for (const trial of transactionsToBeVerified) {
       const encoded = bnsCodec.bytesToPost(trial);
       const decoded = bnsCodec.parseBytes(encoded, trial.transaction.creator.chainId);
-      expect(decoded).toEqual(trial);
+      expect(decoded)
+        .withContext(trial.transaction.kind)
+        .toEqual(trial);
     }
   });
 });