+++ packages/iov-bns/src/encode.ts
@@ -7,7 +7,6 @@ import {
   SignatureBytes,
   SignedTransaction,
   SwapClaimTransaction,
-  SwapCounterTransaction,
   SwapOfferTransaction,
   SwapTimeoutTransaction,
   UnsignedTransaction,
@@ -23,7 +22,7 @@ import {
   RegisterUsernameTx,
   RemoveAddressFromUsernameTx,
 } from "./types";
-import { decodeBnsAddress, identityToAddress, preimageIdentifier } from "./util";
+import { decodeBnsAddress, hashIdentifier, identityToAddress } from "./util";
 
 const { toUtf8 } = Encoding;
 
@@ -120,8 +119,6 @@ export function buildMsg(tx: UnsignedTransaction): codecImpl.app.ITx {
       return buildSendTransaction(tx);
     case "bcp/swap_offer":
       return buildSwapOfferTx(tx);
-    case "bcp/swap_counter":
-      return buildSwapCounterTx(tx);
     case "bcp/swap_claim":
       return buildSwapClaimTx(tx);
     case "bcp/swap_timeout":
@@ -162,22 +159,10 @@ function buildSendTransaction(tx: SendTransaction): codecImpl.app.ITx {
 }
 
 function buildSwapOfferTx(tx: SwapOfferTransaction): codecImpl.app.ITx {
-  const hashed: SwapCounterTransaction = {
-    kind: "bcp/swap_counter",
-    creator: tx.creator,
-    recipient: tx.recipient,
-    amounts: tx.amounts,
-    timeout: tx.timeout,
-    hashCode: preimageIdentifier(tx.preimage),
-  };
-  return buildSwapCounterTx(hashed);
-}
-
-function buildSwapCounterTx(tx: SwapCounterTransaction): codecImpl.app.ITx {
   return {
     createEscrowMsg: codecImpl.escrow.CreateEscrowMsg.create({
       src: decodeBnsAddress(identityToAddress(tx.creator)).data,
-      arbiter: tx.hashCode,
+      arbiter: hashIdentifier(tx.hash),
       recipient: decodeBnsAddress(tx.recipient).data,
       amount: tx.amounts.map(encodeAmount),
       timeout: tx.timeout,